requirements.txt
# 音频分割工具依赖包
# AI-SUMMARY: 音频分割工具所需的Python依赖包列表

# 音频处理核心库
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0

# 音频格式处理
pydub>=0.25.0
soundfile>=0.12.0

# 机器学习
scikit-learn>=1.0.0

# 语音处理
webrtcvad>=2.0.10
SpeechRecognition>=3.10.0

# 中文文本处理
jieba>=0.42.0

# 配置文件处理
PyYAML>=6.0

# 音频降噪
noisereduce>=3.0.0

# 其他工具库
pathlib2>=2.3.0
